moved initial db load (store -> schema).
